Split up topics actions into different files.
