+++ app/modules/topics/actions/apiSaga/apiDelete.js
@@ -1,0 +1,16 @@
+// @flow
+
+import * as a from '../../actionTypes';
+
+const apiDelete = (
+  id: string,
+): a.ApiDeleteTopicAction => {
+  return {
+    type: a.API_DELETE,
+    payload: {
+      id,
+    },
+  };
+};
+
+export default apiDelete;
