+++ app/modules/topics/actions/index.js
@@ -1,0 +1,13 @@
+// @flow
+
+import reducerActions from './reducer';
+import apiSagaActions from './apiSaga';
+import taskSagaActions from './taskSaga';
+
+const actions = {
+  ...reducerActions,
+  ...apiSagaActions,
+  ...taskSagaActions,
+};
+
+export default actions;
