+++ app/modules/topics/actions/reducer/index.js
@@ -1,0 +1,15 @@
+// @flow
+
+import addToState from './addToState';
+import editInState from './editInState';
+import removeFromState from './removeFromState';
+import setItemsInState from './setItemsInState';
+
+const reducerActions = {
+  addToState,
+  editInState,
+  removeFromState,
+  setItemsInState,
+};
+
+export default reducerActions;
