+++ app/modules/topics/actions/reducer/removeFromState.js
@@ -1,0 +1,16 @@
+// @flow
+
+import * as a from '../../actionTypes';
+
+const removeFromState = (
+  id: string,
+): a.RemoveFromStateAction => {
+  return {
+    type: a.REMOVE_FROM_STATE,
+    payload: {
+      id,
+    },
+  };
+};
+
+export default removeFromState;
