+++ app/modules/topics/actions/taskSaga/get.js
@@ -1,0 +1,16 @@
+// @flow
+
+import * as a from '../../actionTypes';
+
+const get = (
+  id: string,
+): a.GetAction => {
+  return {
+    type: a.GET,
+    payload: {
+      id,
+    },
+  };
+};
+
+export default get;
