+++ app/modules/topics/actions/taskSaga/load.js
@@ -1,0 +1,16 @@
+// @flow
+
+import * as a from '../../actionTypes';
+
+const load = (
+  id: string,
+): a.LoadContentAction => {
+  return {
+    type: a.LOAD,
+    payload: {
+      id,
+    },
+  };
+};
+
+export default load;
