+++ app/modules/topics/actions/taskSaga/remove.js
@@ -1,0 +1,16 @@
+// @flow
+
+import * as a from '../../actionTypes';
+
+const remove = (
+  id: string,
+): a.RemoveAction => {
+  return {
+    type: a.REMOVE,
+    payload: {
+      id,
+    },
+  };
+};
+
+export default remove;
