+++ app/modules/topics/saga/task/edit.js
@@ -3,8 +3,8 @@
 import { type Saga } from 'redux-saga';
 import { put } from 'redux-saga/effects';
 
+import actions from '../../actions';
 import * as a from '../../actionTypes';
-import { editInState } from '../../actions';
 
 // eslint-disable-next-line require-yield
 const editSaga = function* (action: a.EditAction): Saga<void> {
@@ -14,7 +14,7 @@ const editSaga = function* (action: a.EditAction): Saga<void> {
     description,
   } = action.payload;
 
-  yield put(editInState(id, title, description));
+  yield put(actions.editInState(id, title, description));
 };
 
 export default editSaga;
