+++ app/modules/topics/saga/task/get.js
@@ -3,11 +3,11 @@
 import { type Saga } from 'redux-saga';
 import { put } from 'redux-saga/effects';
 
+import actions from '../../actions';
 import * as a from '../../actionTypes';
-import { apiGet } from '../../actions';
 
 const getSaga = function* (action: a.GetAction): Saga<void> {
-  yield put(apiGet(action.payload.id));
+  yield put(actions.apiGet(action.payload.id));
 };
 
 export default getSaga;
