+++ app/modules/topics/saga/task/getAll.js
@@ -3,11 +3,11 @@
 import { type Saga } from 'redux-saga';
 import { put } from 'redux-saga/effects';
 
+import actions from '../../actions';
 import * as a from '../../actionTypes';
-import { apiGetAllByUserId } from '../../actions';
 
 const getAllSaga = function* (action: a.GetAllByUserIdAction): Saga<void> {
-  yield put(apiGetAllByUserId(action.payload.userId));
+  yield put(actions.apiGetAllByUserId(action.payload.userId));
 };
 
 export default getAllSaga;
