+++ app/modules/topics/saga/task/load.js
@@ -3,16 +3,16 @@
 import { type Saga } from 'redux-saga';
 import { put } from 'redux-saga/effects';
 
+import actions from '../../actions';
 import * as a from '../../actionTypes';
-import { apiGet, apiGetContent } from '../../actions';
 
 const loadSaga = function* (action: a.LoadContentAction): Saga<void> {
   const {
     id,
   } = action.payload;
 
-  yield put(apiGet(id));
-  yield put(apiGetContent(id));
+  yield put(actions.apiGet(id));
+  yield put(actions.apiGetContent(id));
 };
 
 export default loadSaga;
