+++ app/modules/topics/saga/task/remove.js
@@ -3,8 +3,8 @@
 import { type Saga } from 'redux-saga';
 import { put } from 'redux-saga/effects';
 
+import actions from '../../actions';
 import * as a from '../../actionTypes';
-import { apiDelete } from '../../actions';
 
 // eslint-disable-next-line require-yield
 const removeSaga = function* (action: a.RemoveAction): Saga<void> {
@@ -12,7 +12,7 @@ const removeSaga = function* (action: a.RemoveAction): Saga<void> {
     id,
   } = action.payload;
 
-  yield put(apiDelete(id));
+  yield put(actions.apiDelete(id));
 };
 
 export default removeSaga;
